### 2021-08-15 #3615 @whedon
[PRE REVIEW]: Bam-readcount - rapid generation of basepair-resolution sequence metrics
**Submitting author:** @chrisamiller (<a href="http://orcid.org/0000-0003-4266-6700">Christopher A. Miller</a>)
**Repository:** <a href="https://github.com/genome/bam-readcount" target ="_blank">https://github.com/genome/bam-readcount</a>
**Version:** 1.0
