### 2021-08-15 #3615 @arfon
@chrisamiller - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.  For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

While you wait, if you want to fix the DOIs (noting that Whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
