+++ refs/references.bib
@@ -1,4 +1,5 @@
 @ARTICLE{Bonfield2021-et,
+  doi      = "10.1093/gigascience/giab007",
   title    = "{HTSlib}: {C} library for reading/writing high-throughput
               sequencing data",
   author   = "Bonfield, James K and Marshall, John and Danecek, Petr and Li,
@@ -38,6 +39,7 @@
 }
 
 @ARTICLE{Xie2014-mm,
+  doi      = "10.1038/nm.3733",
   title    = "Age-related mutations associated with clonal hematopoietic
               expansion and malignancies",
   author   = "Xie, Mingchao and Lu, Charles and Wang, Jiayin and McLellan,
@@ -78,6 +80,7 @@
 }
 
 @ARTICLE{Sun2020-fm,
+  doi      = "10.1111/gcb.15291",
   title    = "Rapid genomic and phenotypic change in response to climate
               warming in a widespread plant invader",
   author   = "Sun, Yan and Bossdorf, Oliver and Grados, Ramon D and Liao,
@@ -123,6 +126,7 @@
 }
 
 @ARTICLE{Anzar2019-vp,
+  doi      = "10.1186/s12920-019-0508-5",
   title    = "{NeoMutate}: an ensemble machine learning framework for the
               prediction of somatic mutations in cancer",
   author   = "Anzar, Irantzu and Sverchkova, Angelina and Stratford, Richard
@@ -199,6 +203,7 @@
 }
 
 @ARTICLE{Griffith2015-wx,
+  doi      = "10.1371/journal.pcbi.1004274",
   title    = "Genome Modeling System: A Knowledge Management Platform for
               Genomics",
   author   = "Griffith, Malachi and Griffith, Obi L and Smith, Scott M and
@@ -249,6 +254,7 @@
 }
 
 @ARTICLE{Kockan2017-to,
+  doi      = "10.1093/bioinformatics/btw536",
   title    = "{SiNVICT}: ultra-sensitive detection of single nucleotide
               variants and indels in circulating tumour {DNA}",
   author   = "Kockan, Can and Hach, Faraz and Sarrafi, Iman and Bell, Robert H
@@ -317,6 +323,7 @@
 }
 
 @ARTICLE{Alberti2018-wn,
+  doi      = "10.1038/s41375-018-0193-y",
   title    = "Discriminating a common somatic {ASXL1} mutation (c.1934dup;
               {p.G646Wfs*12}) from artifact in myeloid malignancies using {NGS}",
   author   = "Alberti, Michael O and Srivatsan, Sridhar Nonavinkere and Shao,
@@ -334,6 +341,7 @@
 }
 
 @ARTICLE{Wyatt2016-si,
+  doi      = "10.1001/jamaoncol.2016.0494",
   title    = "Genomic Alterations in {Cell-Free} {DNA} and Enzalutamide
               Resistance in {Castration-Resistant} Prostate Cancer",
   author   = "Wyatt, Alexander W and Azad, Arun A and Volik, Stanislav V and
@@ -402,6 +410,7 @@
 }
 
 @ARTICLE{Griffith2015-gz,
+  doi      = "10.1016/j.cels.2015.08.015",
   title    = "Optimizing cancer genome sequencing and analysis",
   author   = "Griffith, Malachi and Miller, Christopher A and Griffith, Obi L
               and Krysiak, Kilannin and Skidmore, Zachary L and Ramu, Avinash
@@ -440,6 +449,7 @@
 }
 
 @ARTICLE{Jensen2017-gp,
+  doi      = "10.1182/blood-2017-03-735654",
   title    = "The {NCI} Genomic Data Commons as an engine for precision
               medicine",
[…]
