### 2021-09-10 #3615 @lpantano
Hi @oliviamr and @iskandr, would you like and have time to review this tool? Thanks!
As a reminder, our review process happens in github, [read more](https://joss.readthedocs.io/en/latest/review_criteria.html). JOSS encourages best practices and helps the reviewer with a list to go over during the process to make it as easier as possible.
