### 2021-09-13 #3722 @whedon
[REVIEW]: Bam-readcount - rapid generation of basepair-resolution sequence metrics
**Submitting author:** @chrisamiller (<a href="http://orcid.org/0000-0003-4266-6700">Christopher A. Miller</a>)
**Repository:** <a href="https://github.com/genome/bam-readcount" target ="_blank">https://github.com/genome/bam-readcount</a>
**Version:** v1.0.1

- [x] #3731

- [x] #3731
