### 2021-09-15 genome/bam-readcount#81 @friedue
[JOSS REVIEW] Minor edits to the github-based documentation
- [ ] extend the commands for the installation

```
git clone https://github.com/genome/bam-readcount
cd bam-readcount
mkdir build
cd build
cmake ..
make
```

- [ ] perhaps link to [VAtools](https://vatools.readthedocs.io/en/latest/)
