### 2021-09-15 genome/bam-readcount#82 @friedue
[JOSS REVIEW] minor edits to the manuscript
Related to: <https://github.com/openjournals/joss-reviews/issues/3722>

- [x] please extend the "statement of need" a bit, or rephrase it so that within the first sentence it is already obvious that `bam-readcount` is not so much just about counting reads, but primarily about collecting metrics that are particularly helpful in judging wether a specific variant call is reliable or not (or maybe that's not even the primary use-case, who knows! :) )
- [x] similarly, the conclusion should state a bit more clearly what precisely the role of `bam-readcount` is in the bespoke pipelines and the use cases that are mentioned there
- [x] lines 37 and 54 should have "its" instead of "it's"
