### 2021-09-15 #3722 @friedue
I've opened a couple of issues in the repo, most of them dealing with relatively minor issues. The main issue IMO is the installation/docker use and the lack of a proper use-case.

<https://github.com/genome/bam-readcount/issues/79>, <https://github.com/genome/bam-readcount/issues/80>, <https://github.com/genome/bam-readcount/issues/81>, <https://github.com/genome/bam-readcount/issues/82>
