### 2021-09-20 genome/bam-readcount#81 @bebatut
It could also be added that `bam-readcount` can be installed via conda and then give some instruction
