### 2021-09-20 genome/bam-readcount#80 @bebatut
I totally agree with @friedue.

It would be also interesting to
- Give some context about the test data and what to expect to see
- Provide some more real case datasets (or at least where to get them) like the ones mentioned in the paper to evaluate the performance
