### 2021-09-20 genome/bam-readcount#83 @bebatut
[JOSS REVIEW] Missing the community contribution guidelines
Related to https://github.com/openjournals/joss-reviews/issues/3722:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Guidelines to seek support are stated in README file but not how to contribute or report issues.
A [CONTRIBUTING.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file could be added in the repository
