### 2021-09-20 genome/bam-readcount#84 @bebatut
[JOSS REVIEW] Any functionality documentation?
Related to https://github.com/openjournals/joss-reviews/issues/3722:

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Is there any functionality documentation? If there is, where can we find it? I could not find anything at first sight
