### 2021-09-20 #3722 @bebatut
I agree with @friedue about the major issues. I added some comments in the already existing issues and also create new ones:
- Functionality documentation: https://github.com/genome/bam-readcount/issues/84
- Community contribution guidelines: https://github.com/genome/bam-readcount/issues/83

A last comment: is the data used to generate the graphs in the paper available somewhere?
