### 2021-09-22 0869ce4
updating statement of need, correcting typos
