+++ refs/references.bib
@@ -924,3 +924,28 @@
   year     =  2018,
   language = "en"
 }
+@ARTICLE{Koboldt2013-us,
+  doi      = "10.1093/bioinformatics/bty316",
+  title    = "Using VarScan 2 for germline variant calling and somatic mutation detection",
+  author   = "Koboldt, Daniel C and Larson, David E and Wilson, Richard K",
+  abstract = "The identification of small sequence variants remains a 
+             challenging but critical step in the analysis of next-generation 
+             sequencing data. Our variant-calling tool, VarScan 2, employs 
+             heuristic and statistic thresholds based on user-defined criteria 
+             to call variants using SAMtools mpileup data as input. Here, we 
+             provide guidelines for generating that input, and describe protocols 
+             for using VarScan 2 to (1) identify germline variants in individual 
+             samples; (2) call somatic mutations, copy-number alterations, and 
+             LOH events in tumor-normal pairs; and (3) identify germline variants, 
+             de novo mutations, and Mendelian inheritance errors in family trios.
+             Further, we describe a strategy for variant filtering that removes 
+             likely false positives associated with common sequencing- and 
+             alignment-related artifacts",
+  journal  = "Current protocols in bioinformatics",
+  volume   = 44,
+  number   = 1,
+  pages    = "15--4",
+  month    = dec,
+  year     = 2013,
+  language = "en"
+}
