+++ CONTRIBUTING.md
@@ -1,0 +1,15 @@
+# Contributing
+
+We welcome improvements, extensions, bug reports and bug fixes to `bam-readcount`!  
+
+## Bug Reports
+To make bug reports most useful, please include:
+- a brief summary of the problem - what did you expect, and what actually happened?
+- precise steps to reproduce the problem, and a minimal bam/cram file on which the steps can be run, if possible.
+
+## Fixes or Feature Additions
+Direct pull requests are just fine - fork the repository, create an new branch, then make a PR against this repo.
+
+Before investing time and effort, it is often best to create an issue in 
+the github repository reporting the bug and/or proposing the change, and we'll be happy to discuss the 
+details and feasibility with you. 
