### 2021-09-30 81889fc
Update with JOSS review suggestions #82
