### 2021-09-30 b263c12
Update with JOSS review suggestions #81
