### 2021-09-30 genome/bam-readcount#81 @apldx
Good suggestion, @friedue, done. @bebatut, also a great idea about conda, but we will need to update our conda recipe for the new version (looks like it's currently at v0.8 in bioconda), so leaving this open for now.
