### 2021-09-30 genome/bam-readcount#84 @apldx
Though the code is divided into a front-end and library, we don't aim to expose the library code as an API. For the command-line program `bam-readcount`, the intro section of the `README.md` hopefully indicates what it does, and then the actual usage is documented under `Usage`. The output is complicated and so has its own section `Output`. Do you have thoughts on what more might be needed? Thanks!
