### 2021-10-08 65de7fd
Switch to docker-bam-readcount for Docker files
