+++ docker/README.md
@@ -1,25 +1,0 @@
-Docker
-======
-
-Docker image based on Ubuntu Focal with `bam-readcount` `v1.0.0` built
-and installed. Note that this does *not* build from `master`, but the
-`v1.0.0` tag. Edit the `Dockerfile` to use `master` or a different branch
-or tag.
-
-
-Resources
----------
-
-https://github.com/genome/bam-readcount  
-`bam-readcount` and documentation
-
-
-Usage
------
-
-`bam-readcount` is installed as
-
-    /usr/bin/bam-readcount
-
-
-
