### 2021-10-15 genome/bam-readcount#79 @friedue
that looks great
