+++ tutorial/README.md
@@ -1,0 +1,431 @@
+# A `bam-readcount` tutorial
+
+On November 26, 2021, [WHO](https://www.who.int/)
+[classified](https://www.who.int/news/item/26-11-2021-classification-of-omicron-(b.1.1.529)-sars-cov-2-variant-of-concern)
+the
+[SARS-CoV-2](https://en.wikipedia.org/wiki/Severe_acute_respiratory_syndrome_coronavirus_2)
+[variant](https://covariants.org/) [`B.1.1.529`](https://covariants.org/variants/21K.Omicron) as a ["Variant of
+Concern"](https://www.who.int/en/activities/tracking-SARS-CoV-2-variants/) (VOC), designated
+"[Omicron](https://en.wikipedia.org/wiki/Omicron)". The variant was
+noted as having an unusually large number of mutations, including more
+than 30 in the
+[Spike protein](https://en.wikipedia.org/wiki/Coronavirus_spike_protein).
+
+On November 29, [Tulio de Oliveira](https://twitter.com/Tuliodna)
+of [CERI](https://ceri.org.za/) [posted](https://twitter.com/Tuliodna/status/1465338678264401934) a
+[link](https://www.ncbi.nlm.nih.gov/sra/?term=PRJNA784038) to the
+[FASTQ](https://en.wikipedia.org/wiki/FASTQ_format) sequencing files for some
+of the first sequenced Omicron cases. We can use
+[`bam-readcount`](https://github.com/genome/bam-readcount) along with the
+SARS-CoV-2 [reference genome](https://www.ncbi.nlm.nih.gov/nuccore/1798174254) and some common software to explore these mutations. Please
+note that this is just an exercise and would surely not be street-legal in a
+virology department!
+
+The only requirement is [Docker](https://www.docker.com/) and access to a Unix
+command line, although all the tools can be easily built locally instead. A 
+ready-to-run Bash script containing all the commands is available 
+[here](scripts/commands.sh)
+
+## Docker setup
+
+All Docker commands use similar arguments to allow us to access files in the
+current working directory, so first we will define an alias with the
+common options
+
+    # --rm remove container on exit
+    # -w $(pwd) set working directory to current working directory
+    # -v $(pwd):$(pwd) map current working directory inside container
+    alias dockercmd='docker run --rm -w $(pwd) -v $(pwd):$(pwd)'
+
+and run Docker as
+
+    dockercmd IMAGE ARGS...
+
+All the computations can be run quickly on an ordinary laptop; no command
+should take more than a minute to run. The Docker containers are also 
+fairly small and should download quickly on a broadband connection.
+
+If you are using a shell that does not support this `alias` format, you 
+can just substitute the full command for `dockercmd` in the command
+lines below.
+
+
+## Download the reference and FASTQs
+
+    # SARS-CoV-2 standard reference genome
+    dockercmd curlimages/curl -o NC_045512.2.fa "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=nuccore&id=NC_045512.2&rettype=fasta"
+
+    # CERI FASTQs via the SRA for case CERI-KRISP-K032274
+    # We will use the SRA tools to download and convert to FASTQ
[…]
