### 2021-12-16 genome/bam-readcount#80 @chrisamiller
This was the biggest thing remaining, and @apldx addressed it with the merge he made yesterday, which adds an extensive (and relevant!) tutorial:  https://github.com/genome/bam-readcount/tree/master/tutorial Marking this done.
