### 2021-12-16 #3722 @chrisamiller
This got delayed by a variety of other projects on our end - appreciate your patience!  At this point we believe that we've addressed all of the the reviewer concerns.  They were really useful comments, and I think the package and the paper are stronger as a result - much appreciated! (@bebatut @friedue)
