### 2021-12-16 genome/bam-readcount#82 @friedue
Great update to the text! The use-cases of the tool are very clearly described now, great job!

[ ] There's a typo on line 61, and there might be a letter missing in line 50, too ("disks?").
