### 2021-12-16 genome/bam-readcount#82 @chrisamiller
Thanks - fixed and pushed!
