### 2021-12-16 genome/bam-readcount#80 @friedue
This is a great and timely tutorial! It ticks all the boxes -- it show cases bam-readcount, it demonstrates what its place could be in typical analysis pipelines and it features some hands-on, highly relevant and scientifically interesting results. Well worth the wait! Now you just need to link/highlight this tutorial in the bam-readcount repo README.
