### 2021-12-16 genome/bam-readcount#80 @chrisamiller
Done!
