### 2021-12-16 genome/bam-readcount#80 @friedue
You could really be capitalizing on the omicron-link to get people to use bam-readcount and familiarize themselves with it without noticing (because they'll be so eager to find out more about omicron).
But that's your call and it's totally fine the way you've done it now. Shout-out to @apldx for a job really well done!
