### 2021-12-17 genome/bam-readcount#80 @apldx
Thanks! It was really fun to work on. Seconding @chrisamiller, really appreciate all the suggestions.
