### 2022-01-11 genome/bam-readcount#83 @bebatut
Looks good. Thanks
