### 2022-01-11 genome/bam-readcount#84 @bebatut
Looks good
