### 2022-01-18 e674b73
Unix conv help message edits
