### 2022-01-19 #3722 @chrisamiller
- Release is v1.0.1
- It has been archived in [Zenodo](https://zenodo.org/record/5879149)  I am unclear on how to make the title match the paper - we have Github/Zenodo set up to auto-archive every tagged version, and it auto-generates that title.  Would including the paper title in the description of the repository be sufficient?
- Metadata all matches up
-  Zenodo DOI: 10.5281/zenodo.5879149

Thanks!
